requirements.txt
communex 
typer
uvicorn
keylimiter
pydantic-settings
numpy
torch
torchaudio
transformers
scipy
pydub